breakout audioController into it's own module
